Fix data types and enhance functionality in CUDA beam search implementation
- Changed z_idx from int to long long in evaluate_rotations_kernel to accommodate larger index values.
- Updated initial grid size from 4 to 24 in main-v1.1.1.cpp for improved problem complexity.
- Removed unnecessary variable in beam_search function and added flexibility in paired count checks.
- Initialized zobrist hashing in main function for better performance.
- Updated match.json with a new problem field containing a larger grid for testing.

+++ cuda_beam_search.cu
@@ -69,7 +69,7 @@ __global__ void evaluate_rotations_kernel(
         max_val = val;
 
       // ZOBRIST_MAX_N=64, ZOBRIST_MAX_VAL=2048
-      int z_idx = (r * 64 * 2048) + (c * 2048) + val;
+      long long z_idx = (r * 64 * 2048) + (c * 2048) + val;
       hash ^= zobrist_table[z_idx];
     }
   }
